requirements.txt
# Core Dependencies
google-genai>=0.3.0
aiohttp>=3.9.0
asyncio
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
websockets>=12.0
python-multipart>=0.0.6

# Database & Storage
sqlalchemy>=2.0.0
cryptography>=41.0.0

# System Monitoring
psutil>=5.9.0

# Utilities
python-dotenv>=1.0.0